classe Plot criada
